requirements.txt
blinker==1.7.0
cachelib==0.12.0
Cerberus==1.3.5
click==8.1.7
Flask==3.0.2
Flask-Session==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
pip==24.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
stringcase==1.2.0
typing_extensions==4.10.0
Werkzeug==3.0.1
